Many changes for Unicode, by Marc-Andre Lemburg.

+++ Objects/intobject.c
@@ -942,7 +942,7 @@ PyInt_Fini()
 				if (PyInt_Check(p) && p->ob_refcnt != 0)
 					fprintf(stderr,
 				"#   <int at %lx, refcnt=%d, val=%ld>\n",
-						p, p->ob_refcnt, p->ob_ival);
+					  (long)p, p->ob_refcnt, p->ob_ival);
 			}
 			list = list->next;
 		}
